Prepared infrastructure for bots
Left it ready for Bjorn to add a call and response system.
The Game Manager script now takes whatever the player typed and assings it a string variable called 'playerInput' which we could feed into the bot. The dialogue script now takes whatever string the bot returns and fits it neatly into it's GUI